# docs/swagger.yaml
basePath: /
definitions:
...
  headless-cms_internal_models.Section:
    properties:
      children:
        items:
          $ref: '#/definitions/headless-cms_internal_models.Section'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parent_id:
        type: integer
      updated_at:
        type: string
    type: object
...
  internal_handler.ErrorResponse:
    properties:
      error:
        example: описание ошибки
        type: string
    type: object
...
info:
  contact: {}
  description: Headless CMS без фронтенда. Разделы, статьи (Markdown), новости (Markdown
    + автогенерация краткого описания), пользователи/роли, медиа (MinIO).
  title: Headless CMS API
  version: "1.0"
paths:
...
  /api/v1/sections:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/headless-cms_internal_models.Section'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Получить все разделы
      tags:
      - Разделы
    post:
      consumes:
      - application/json
      parameters:
      - description: Раздел (id игнорируется)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/headless-cms_internal_models.Section'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/headless-cms_internal_models.Section'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать раздел
      tags:
      - Разделы
  /api/v1/sections/{id}:
    delete:
      parameters:
      - description: ID раздела
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить раздел
      tags:
      - Разделы
    get:
      parameters:
      - description: ID раздела
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/headless-cms_internal_models.Section'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Получить раздел по ID
      tags:
      - Разделы
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID раздела
        in: path
        name: id
        required: true
        type: integer
      - description: Раздел (id из path)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/headless-cms_internal_models.Section'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/headless-cms_internal_models.Section'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить раздел
      tags:
      - Разделы
...
  /api/v1/sections/{id}/children:
    get:
      parameters:
      - description: ID родительского раздела
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/headless-cms_internal_models.Section'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_handler.ErrorResponse'
      summary: Получить дочерние разделы
      tags:
      - Разделы
...
securityDefinitions:
  BearerAuth:
    description: 'Введите JWT токен в формате: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"